+++ src/components/app/EnvelopesWrapper.tsx
@@ -3,7 +3,6 @@ import * as messages from '@cucumber/messages'
 import { Query as CucumberQuery } from '@cucumber/query'
 import React, { FunctionComponent, PropsWithChildren, useMemo } from 'react'
 
-import { EnvelopesQuery } from '../../EnvelopesQueryContext.js'
 import { QueriesWrapper } from './QueriesWrapper.js'
 
 interface IProps {
@@ -14,23 +13,17 @@ export const EnvelopesWrapper: FunctionComponent<PropsWithChildren<IProps>> = ({
   envelopes,
   children,
 }) => {
-  const { gherkinQuery, cucumberQuery, envelopesQuery } = useMemo(() => {
+  const { gherkinQuery, cucumberQuery } = useMemo(() => {
     const gherkinQuery = new GherkinQuery()
     const cucumberQuery = new CucumberQuery()
-    const envelopesQuery = new EnvelopesQuery()
     for (const envelope of envelopes) {
       gherkinQuery.update(envelope)
       cucumberQuery.update(envelope)
-      envelopesQuery.update(envelope)
     }
-    return { gherkinQuery, cucumberQuery, envelopesQuery }
+    return { gherkinQuery, cucumberQuery }
   }, [envelopes])
   return (
-    <QueriesWrapper
-      gherkinQuery={gherkinQuery}
-      cucumberQuery={cucumberQuery}
-      envelopesQuery={envelopesQuery}
-    >
+    <QueriesWrapper gherkinQuery={gherkinQuery} cucumberQuery={cucumberQuery}>
       {children}
     </QueriesWrapper>
   )