+++ src/components/app/FilteredResults.tsx
@@ -1,67 +1,22 @@
 import React, { FC } from 'react'
 
-import { useQueries, useSearch } from '../../hooks/index.js'
-import { useFilteredDocuments } from '../../hooks/useFilteredDocuments.js'
-import { useResultStatistics } from '../../hooks/useResultStatistics.js'
 import { ExecutionSummary } from './ExecutionSummary.js'
+import { FilteredDocuments } from './FilteredDocuments.js'
 import styles from './FilteredResults.module.scss'
-import { FilteredResultsExperimental } from './FilteredResultsExperimental.js'
-import { GherkinDocumentList } from './GherkinDocumentList.js'
-import { NoMatchResult } from './NoMatchResult.js'
 import { SearchBar } from './SearchBar.js'
 import { StatusesSummary } from './StatusesSummary.js'
 
-interface Props {
+export const FilteredResults: FC<{
   className?: string
-  experimental?: boolean
-}
-
-export const FilteredResults: FC<Props> = ({ className, experimental }) => {
-  return experimental ? (
-    <FilteredResultsExperimental className={className} />
-  ) : (
-    <FilteredResultsCurrent className={className} />
-  )
-}
-
-const FilteredResultsCurrent: FC<Omit<Props, 'experimental'>> = ({ className }) => {
-  const { envelopesQuery } = useQueries()
-  const { scenarioCountByStatus, statusesWithScenarios, totalScenarioCount } = useResultStatistics()
-  const { query, hideStatuses, update } = useSearch()
-  const filtered = useFilteredDocuments(query, hideStatuses)
-
+}> = ({ className }) => {
   return (
     <div className={className}>
       <div className={styles.reportHeader}>
-        <StatusesSummary
-          scenarioCountByStatus={scenarioCountByStatus}
-          totalScenarioCount={totalScenarioCount}
-        />
-        <ExecutionSummary
-          scenarioCountByStatus={scenarioCountByStatus}
-          totalScenarioCount={totalScenarioCount}
-          testRunStarted={envelopesQuery.getTestRunStarted()}
-          testRunFinished={envelopesQuery.getTestRunFinished()}
-          meta={envelopesQuery.getMeta()}
-        />
-        <SearchBar
-          query={query}
-          onSearch={(newValue) => update({ query: newValue })}
-          statusesWithScenarios={statusesWithScenarios}
-          hideStatuses={hideStatuses}
-          onFilter={(newValue) => update({ hideStatuses: newValue })}
-        />
+        <StatusesSummary />
+        <ExecutionSummary />
+        <SearchBar />
       </div>
-
-      {filtered !== undefined && (
-        <>
-          {filtered.length > 0 ? (
-            <GherkinDocumentList gherkinDocuments={filtered} preExpand={true} />
-          ) : (
-            <NoMatchResult query={query} />
-          )}
-        </>
-      )}
+      <FilteredDocuments />
     </div>
   )
 }