+++ src/components/app/StatusesSummary.spec.tsx
@@ -1,31 +1,23 @@
-import { TestStepResultStatus } from '@cucumber/messages'
 import { render } from '@testing-library/react'
 import { expect } from 'chai'
 import React from 'react'
 
-import { makeEmptyScenarioCountsByStatus } from '../../countScenariosByStatuses.js'
+import examplesTablesFeature from '../../../acceptance/examples-tables/examples-tables.feature.js'
+import { EnvelopesWrapper } from './EnvelopesWrapper.js'
 import { StatusesSummary } from './StatusesSummary.js'
 
 describe('StatusesSummary', () => {
   it('should render correctly', () => {
     const { getAllByRole } = render(
-      <StatusesSummary
-        scenarioCountByStatus={{
-          ...makeEmptyScenarioCountsByStatus(),
-          ...{
-            [TestStepResultStatus.PASSED]: 100,
-            [TestStepResultStatus.FAILED]: 3,
-            [TestStepResultStatus.UNDEFINED]: 1,
-          },
-        }}
-        totalScenarioCount={104}
-      />
+      <EnvelopesWrapper envelopes={examplesTablesFeature}>
+        <StatusesSummary />
+      </EnvelopesWrapper>
     )
 
     expect(getAllByRole('listitem').map((li) => li.textContent)).to.deep.eq([
-      '3 failed',
-      '100 passed',
-      '1 undefined',
+      '2 failed',
+      '5 passed',
+      '2 undefined',
     ])
   })
 })