+++ src/components/app/StatusesSummary.stories.tsx
@@ -1,48 +1,31 @@
-import { TestStepResultStatus } from '@cucumber/messages'
+import * as messages from '@cucumber/messages'
 import { Story } from '@ladle/react'
 import React from 'react'
 
-import { makeEmptyScenarioCountsByStatus } from '../../countScenariosByStatuses.js'
+import examplesTablesFeature from '../../../acceptance/examples-tables/examples-tables.feature.js'
 import { CucumberReact } from '../CucumberReact.js'
-import { IStatusesSummaryProps, StatusesSummary } from './StatusesSummary.js'
+import { EnvelopesWrapper } from './EnvelopesWrapper.js'
+import { StatusesSummary } from './StatusesSummary.js'
 
 export default {
   title: 'App/StatusesSummary',
 }
 
-const Template: Story<IStatusesSummaryProps> = (props) => {
+type TemplateArgs = {
+  envelopes: readonly messages.Envelope[]
+}
+
+const Template: Story<TemplateArgs> = ({ envelopes }) => {
   return (
     <CucumberReact>
-      <StatusesSummary {...props} />
+      <EnvelopesWrapper envelopes={envelopes}>
+        <StatusesSummary />
+      </EnvelopesWrapper>
     </CucumberReact>
   )
 }
 
-const scenarioCountByStatus = {
-  ...makeEmptyScenarioCountsByStatus(),
-  ...{
-    [TestStepResultStatus.PASSED]: 100,
-    [TestStepResultStatus.FAILED]: 3,
-    [TestStepResultStatus.UNDEFINED]: 1,
-  },
-}
-
 export const Typical = Template.bind({})
 Typical.args = {
-  scenarioCountByStatus,
-  totalScenarioCount: 104,
-} as IStatusesSummaryProps
-
-export const All = Template.bind({})
-All.args = {
-  scenarioCountByStatus: {
-    [TestStepResultStatus.PASSED]: 3,
-    [TestStepResultStatus.FAILED]: 3,
-    [TestStepResultStatus.PENDING]: 3,
-    [TestStepResultStatus.SKIPPED]: 3,
-    [TestStepResultStatus.UNDEFINED]: 3,
-    [TestStepResultStatus.AMBIGUOUS]: 3,
-    [TestStepResultStatus.UNKNOWN]: 3,
-  },
-  totalScenarioCount: 21,
-} as IStatusesSummaryProps
+  envelopes: examplesTablesFeature,
+} as TemplateArgs