+++ src/filter/filterByStatus.ts
@@ -1,16 +1,13 @@
 import { Query as GherkinQuery } from '@cucumber/gherkin-utils'
 import { GherkinDocumentWalker, rejectAllFilters } from '@cucumber/gherkin-utils'
 import * as messages from '@cucumber/messages'
-import { getWorstTestStepResult } from '@cucumber/messages'
+import { getWorstTestStepResult, TestCase } from '@cucumber/messages'
 import { Query as CucumberQuery } from '@cucumber/query'
-
-import { EnvelopesQuery } from '../EnvelopesQueryContext.js'
 
 export default function filterByStatus(
   gherkinDocument: messages.GherkinDocument,
   gherkinQuery: GherkinQuery,
   cucumberQuery: CucumberQuery,
-  envelopesQuery: EnvelopesQuery,
   statuses: readonly messages.TestStepResultStatus[]
 ): messages.GherkinDocument | null {
   const filters = {
@@ -18,8 +15,12 @@ export default function filterByStatus(
       if (!gherkinDocument.uri) throw new Error('Missing uri for gherkinDocument')
       const pickleIds = gherkinQuery.getPickleIds(gherkinDocument.uri, scenario.id)
 
+      const picklesIdsWithTestCases = cucumberQuery
+        .findAllTestCaseStarted()
+        .map((tcs) => (cucumberQuery.findTestCaseBy(tcs) as TestCase).pickleId)
+
       return pickleIds
-        .filter((pickleId) => envelopesQuery.hasTestCase(pickleId))
+        .filter((pickleId) => picklesIdsWithTestCases.includes(pickleId))
         .some((pickleId) =>
           statuses.includes(
             getWorstTestStepResult(cucumberQuery.getPickleTestStepResults([pickleId])).status