+++ src/hooks/useFilteredDocuments.ts
@@ -1,16 +1,15 @@
-import { GherkinDocument, TestStepResultStatus } from '@cucumber/messages'
+import { GherkinDocument } from '@cucumber/messages'
 import { useEffect, useState } from 'react'
 
 import { allStatuses } from '../countScenariosByStatuses.js'
 import filterByStatus from '../filter/filterByStatus.js'
 import { createSearch, Searchable } from '../search/index.js'
 import { useQueries } from './useQueries.js'
+import { useSearch } from './useSearch.js'
 
-export function useFilteredDocuments(
-  query: string,
-  hideStatuses: readonly TestStepResultStatus[]
-): GherkinDocument[] | undefined {
-  const { gherkinQuery, cucumberQuery, envelopesQuery } = useQueries()
+export function useFilteredDocuments(): GherkinDocument[] | undefined {
+  const { query, hideStatuses } = useSearch()
+  const { gherkinQuery, cucumberQuery } = useQueries()
   const [searchable, setSearchable] = useState<Searchable>()
   const [results, setResults] = useState<GherkinDocument[]>()
   useEffect(() => {
@@ -28,7 +27,6 @@ export function useFilteredDocuments(
               document,
               gherkinQuery,
               cucumberQuery,
-              envelopesQuery,
               allStatuses.filter((s) => !hideStatuses.includes(s))
             )
           )
@@ -36,6 +34,6 @@ export function useFilteredDocuments(
         setResults(filtered)
       }
     )
-  }, [query, hideStatuses, gherkinQuery, cucumberQuery, envelopesQuery, searchable])
+  }, [query, hideStatuses, gherkinQuery, cucumberQuery, searchable])
   return results
 }